feat: modularize implementation_service/core.py into focused submodules
Split the 2856-line core.py monolith into cohesive submodules, a pure
behavior-preserving mechanical move (no logic changes):

- draft_pr.py   — bootstrap_draft_pr, build_implementation_prompt, plan markers
- sync.py       — sync, catchup, classify_staleness, _merge_base, preflight
- lifecycle.py  — post-implementation lifecycle, merge, PR-body composition
- done.py       — done, _done_via_pr, _done_via_direct
- cleanup.py    — list_sessions, remove, _preserve_session_data, _cleanup_worktree
- _shared.py    — find_worktree_path, extract_issue_from_branch (breaks the
                  core->lifecycle->cleanup->core import cycle)
- core.py       — slimmed to start + task/worktree resolution + usage capture
                  (2869 -> 775 lines)

The package __init__.py re-exports every public name so existing imports
(from wade.services.implementation_service import start|sync|done|...) are
unchanged. Test patch targets are repointed to each moved function's new
submodule (patched where the name is looked up).

Closes #314
