+++ src/wade/services/implementation_service/__init__.py
@@ -5,7 +5,13 @@ Existing imports like ``from wade.services.implementation_service import start``
 continue to work without changes.
 """
 
+from wade.services.implementation_service._shared import *  # noqa: F403
 from wade.services.implementation_service.batch import *  # noqa: F403
 from wade.services.implementation_service.bootstrap import *  # noqa: F403
+from wade.services.implementation_service.cleanup import *  # noqa: F403
 from wade.services.implementation_service.core import *  # noqa: F403
+from wade.services.implementation_service.done import *  # noqa: F403
+from wade.services.implementation_service.draft_pr import *  # noqa: F403
+from wade.services.implementation_service.lifecycle import *  # noqa: F403
+from wade.services.implementation_service.sync import *  # noqa: F403
 from wade.services.implementation_service.usage_tracking import *  # noqa: F403
