+++ tests/unit/test_services/test_done_plan_flag.py
@@ -84,29 +84,29 @@ def test_done_with_plan_flag_resolves_and_delegates() -> None:
     provider = MagicMock()
     with (
         patch(
-            "wade.services.implementation_service.core.load_config",
+            "wade.services.implementation_service.done.load_config",
             return_value=ProjectConfig(project=ProjectSettings(main_branch="main")),
         ),
-        patch("wade.services.implementation_service.core.get_provider", return_value=provider),
+        patch("wade.services.implementation_service.done.get_provider", return_value=provider),
         patch(
-            "wade.services.implementation_service.core.git_repo.get_repo_root",
+            "wade.services.implementation_service.done.git_repo.get_repo_root",
             return_value=Path("/tmp/repo"),
         ),
         patch(
-            "wade.services.implementation_service.core._resolve_worktree_from_plan",
+            "wade.services.implementation_service.done._resolve_worktree_from_plan",
             return_value=(Path("/tmp/wt"), "feat/42-my-plan", "42"),
         ) as mock_resolve,
         patch(
-            "wade.services.implementation_service.core.find_worktree_path",
+            "wade.services.implementation_service.done.find_worktree_path",
             return_value=Path("/tmp/wt"),
         ),
         patch(
-            "wade.services.implementation_service.core.git_repo.get_current_branch",
+            "wade.services.implementation_service.done.git_repo.get_current_branch",
             return_value="feat/42-my-plan",
         ),
-        patch("wade.services.implementation_service.core.git_repo.is_clean", return_value=True),
+        patch("wade.services.implementation_service.done.git_repo.is_clean", return_value=True),
         patch(
-            "wade.services.implementation_service.core._done_via_pr", return_value=True
+            "wade.services.implementation_service.done._done_via_pr", return_value=True
         ) as mock_done,
     ):
         result = implementation_service.done(plan_file=Path("PLAN.md"))
@@ -121,15 +121,15 @@ def test_done_plan_flag_error_returns_false() -> None:
 
     with (
         patch(
-            "wade.services.implementation_service.core.load_config", return_value=ProjectConfig()
+            "wade.services.implementation_service.done.load_config", return_value=ProjectConfig()
         ),
-        patch("wade.services.implementation_service.core.get_provider", return_value=MagicMock()),
+        patch("wade.services.implementation_service.done.get_provider", return_value=MagicMock()),
         patch(
-            "wade.services.implementation_service.core.git_repo.get_repo_root",
+            "wade.services.implementation_service.done.git_repo.get_repo_root",
             return_value=Path("/tmp/repo"),
         ),
         patch(
-            "wade.services.implementation_service.core._resolve_worktree_from_plan",
+            "wade.services.implementation_service.done._resolve_worktree_from_plan",
             side_effect=ValueError("No worktree found"),
         ),
     ):
